- double

+++ op/conv2d_gradfix.py
@@ -135,11 +135,11 @@ def conv2d_gradfix(
         @staticmethod
         def forward(ctx, input, weight, bias):
             if not transpose:
-                out = F.conv2d(input=input.double(), weight=weight.double(), bias=bias, **common_kwargs)
+                out = F.conv2d(input=input, weight=weight, bias=bias, **common_kwargs)
 
             else:
                 out = F.conv_transpose2d(
-                    input=input.double(),
+                    input=input,
                     weight=weight,
                     bias=bias,
                     output_padding=output_padding,
